+++ lib/Listener/LoadAdditionalScriptsListener.php
@@ -2,9 +2,9 @@
 
 namespace OCA\OpenProject\Listener;
 
-use OC_Util;
 use OCA\Files\Event\LoadAdditionalScriptsEvent;
 use OCA\OpenProject\AppInfo\Application;
+use OCA\OpenProject\ServerVersionHelper;
 use OCP\EventDispatcher\Event;
 use OCP\EventDispatcher\IEventListener;
 use OCP\Util;
@@ -13,15 +13,12 @@ use OCP\Util;
  *	@template-implements IEventListener<Event>
  */
 class LoadAdditionalScriptsListener implements IEventListener {
-	public function __construct() {
-	}
 
 	public function handle(Event $event): void {
 		if (!$event instanceof LoadAdditionalScriptsEvent) {
 			return;
 		}
-
-		if (version_compare(implode('.', OC_Util::getVersion()), '28') < 0) {
+		if (version_compare(ServerVersionHelper::getNextcloudVersion(), '28') < 0) {
 			Util::addScript(Application::APP_ID, Application::APP_ID . '-fileActions');
 			Util::addScript(Application::APP_ID, Application::APP_ID . '-filesPluginLessThan28', 'files');
 		} else {